Norminette Fixed Fully

+++ raycasting/line_drawer.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   line_drawer.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: afaqir <afaqir@student.42.fr>              +#+  +:+       +#+        */
+/*   By: agoujdam <agoujdam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/15 05:35:14 by afaqir            #+#    #+#             */
-/*   Updated: 2024/03/15 05:41:33 by afaqir           ###   ########.fr       */
+/*   Updated: 2024/03/16 20:15:32 by agoujdam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	my_mlx_pixel_put(t_img img, int x, int y, unsigned int color)
 	}
 }
 
-unsigned int	rgb_to_hex(int r, int g, int b)
+size_t	rgb_to_hex(int r, int g, int b)
 {
 	return ((r << 16) | (g << 8) | b);
 }
